+++ src/app/(board)/online/viewer/[game_id]/_components/ViewerBoard/ViewerBoard.tsx
