+++ src/app/(board)/online/viewer/[game_id]/_components/ViewerHeader/ViewerHeader.tsx
