+++ src/app/(default)/online/games/_components/GameListTable/GameListTable.tsx
@@ -5,7 +5,6 @@ import { IconAdjustmentsHorizontal } from "@tabler/icons-react";
 import { cdate } from "cdate";
 
 import PublicityBadge from "../PublicityBadge/PublicityBadge";
-import ShareGameButton from "../ShareGameButton/ShareGameButton";
 
 import ButtonLink from "@/app/_components/ButtonLink";
 import Link from "@/app/_components/Link";
@@ -60,10 +59,6 @@ const GameListTable: React.FC<GameListTableProps> = ({ gameList }) => {
                   </Table.Td>
                   <Table.Td>
                     <Group gap="xs">
-                      <ShareGameButton
-                        gameId={game.id}
-                        isPublic={game.isPublic}
-                      />
                       <ButtonLink
                         href={`/online/games/${game.id}/config`}
                         leftSection={<IconAdjustmentsHorizontal />}
